% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_power.R
\name{aggregate_mvoslr_power_object}
\alias{aggregate_mvoslr_power_object}
\title{Aggregates the results of several simulations. Before the aggregation itself, it is checked whether the objects are compatible in terms of the varied parameter and its values.}
\usage{
aggregate_mvoslr_power_object(power_obj_list)
}
\arguments{
\item{power_obj_list}{List of objects of class "mvoslr_power_object"}
}
\value{
Object of type "mvoslr_power_object" in which the information of the objects in the list was aggregated
}
\description{
Aggregates the results of several simulations. Before the aggregation itself, it is checked whether the objects are compatible in terms of the varied parameter and its values.
}
\keyword{internal}
