% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discretize_functions}
\alias{discretize_functions}
\title{Discretize list of functions for given time window and given number of steps}
\usage{
discretize_functions(function_list, max_argument, time_steps = 100)
}
\arguments{
\item{function_list}{list of functions that shall be discretized}

\item{max_argument}{size of time window}

\item{time_steps}{number of steps}
}
\value{
Data frame containing values of functions at chosen input values
}
\description{
Discretize list of functions for given time window and given number of steps
}
\examples{
f1 <- function(t) t^1.1
f2 <- function(t) t^0.9
f3 <- function(t) t^1.2
function_list_example <- list(f1, f2, f3)
mvoslr:::discretize_functions(function_list_example, max_argument = 10, time_steps = 1000)
}
\keyword{internal}
