% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_mvoslr.R
\name{execution_mvoslr}
\alias{execution_mvoslr}
\title{Title}
\usage{
execution_mvoslr(
  msm_data,
  analysis_dates,
  current_analysis = NULL,
  transition_matrix,
  cum_hazard_functions,
  model_type,
  events,
  norm = "l2",
  boundaries = "obf",
  alpha = 0.05,
  weights = NULL
)
}
\arguments{
\item{msm_data}{data frame containing}

\item{analysis_dates}{vector of calendar dates of analyses}

\item{current_analysis}{number of the current analysis}

\item{transition_matrix}{matrix of transitions between states as in mstate package}

\item{cum_hazard_functions}{cumulative hazard functions for transitions in this model}

\item{events}{list of}

\item{norm}{use either \eqn{L^2}-norm (\code{norm = "l2"}) or \eqn{L^\infty}-norm (\code{norm = "linf"}) of vector of test statistics to compute stagewise p-values}

\item{boundaries}{use either}

\item{weights}{}
}
\value{

}
\description{
Title
}
