% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_mvoslr_fixed_fu.R
\name{execution_mvoslr_fixed_fu}
\alias{execution_mvoslr_fixed_fu}
\title{Execution of interim and final analyses of multivariate group-sequential one-sample log-rank test with varying accrual duration
and a fixed follow-up after end of the accrual period}
\usage{
execution_mvoslr_fixed_fu(
  msm_data,
  interim_analysis_dates,
  follow_up,
  current_analysis = NULL,
  transition_matrix,
  cum_hazard_functions,
  model_type,
  events,
  accrual_durations,
  norm = "l2",
  boundaries = "obf",
  alpha = 0.05,
  weights = NULL
)
}
\arguments{
\item{msm_data}{Data frame containing information about any (possible) transition of patients in the study with the following columns:
\itemize{
\item id - ID of the patient to whom this entry belongs
\item Tstart - Start of the observation period for this transition
\item Tstop - Stop of the observation period for this transition
\item duration - Duration of the observation period for this transition
\item from - Current state of the patient
\item to - State to which a transition is possible
\item status - Transition could (1) or could not be (0) observed
\item trans - Number of this transition (according to \code{to.trans2(transition_matrix)})
\item recruitment_date - Calendar date of recruitment of this patient
}
Maximum number of sample sizes under consideration should be at most the number of subjects in this data set}

\item{interim_analysis_dates}{Vector of calendar dates of interim analyses}

\item{follow_up}{Duration of the follow-up period (after stop of recruitment)}

\item{current_analysis}{Number of the current analysis}

\item{transition_matrix}{Matrix of transitions between states as in mstate package}

\item{cum_hazard_functions}{Cumulative hazard functions for transitions in this model}

\item{model_type}{Reference multi-state model is either Markov (\code{model_type = "M"}) or Semi-Markov (\code{model_type = "SM"})}

\item{events}{List of (composite) events that shall be investigated}

\item{accrual_durations}{Vector of durations of the accrual period that shall be investigated}

\item{norm}{Use either \eqn{L^2}-norm (\code{norm = "l2"}, default value) or \eqn{L^\infty}-norm (\code{norm = "linf"}) of vector of test statistics to compute stagewise p-values}

\item{boundaries}{Use either O'Brien-Fleming'S (\code{boundaries = "obf"}, default value) or Pocock's (\code{boundaries = "pocock"}) sequential decision boundaries}

\item{alpha}{Choose type I error rate (default value = 0.05)}

\item{weights}{Choose weights for inverse normal combination of stagewise p-values. Sum of squared values needs to sum up to 1.}
}
\value{
List of 4:
\itemize{
\item raw_martingale - Values of the multivariate process at all stages for all sample sizes
\item stagewise_p_values - Stagewise p-values for all sample sizes
\item decision - Decision to be made at the current analysis for each sample size
\item decision_stage - Number of analysis at which the null hypothesis is rejected (\code{NA} if not applicable)
for each sample size
}
}
\description{
For a given data set, this function can efficiently compute the results of the group-sequential multivariate one-sample
log-rank test for different durations of the accrual duration. Calendar dates of interim_analyses are fixed and the calendar
date of the final analysis is set to the duration of the follow-up period plus the duration of the respective accrual period.
This function is needed to efficiently estimate the power of such trials for different sample sizes via simulation.
}
