% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tihmm.R
\name{fit_tihmm}
\alias{fit_tihmm}
\title{Fitting of time-inhomogeneous Markov model with Weibull transition intensities to (historic) data from multi-state model}
\usage{
fit_tihmm(msm_data, transition_matrix, joint_shape = TRUE)
}
\arguments{
\item{msm_data}{data frame containing information about any (possible) transition of patients in the study with the following columns:
\itemize{
\item id - ID of the patient to whom this entry belongs
\item Tstart - Start of the observation period for this transition
\item Tstop - Stop of the observation period for this transition
\item duration - Duration of the observation period for this transition
\item from - Current state of the patient
\item to - State to which a transition is possible
\item status - Transition could (1) or could not be (0) observed
\item trans - Number of this transition (according to \code{to.trans2(transition_matrix)})
}}

\item{transition_matrix}{matrix of transitions between states as in mstate package}

\item{joint_shape}{boolean variable indicating whether shape parameter is the same for all transitions or not}
}
\value{
List of 2:
\itemize{
\item parameters - Estimated parameters for each transition
\item cum_hazard_function - List of cumulative hazard function to be passed to power analysis or execution of multivariate one-sample log-rank test
}
}
\description{
Fitting of time-inhomogeneous Markov model with Weibull transition intensities to (historic) data from multi-state model
}
\examples{
#Setup of reference multi-state model (here: simple illness-death model)
library(mstate)
tmat_example <- transMat(x = list(c(2,3),c(3),c()), names = c("a", "b", "c"))
#Fictional data from observations in multi-state model with one dead patient without illness
#  (\code{id =1}), one dead patient with illness (\code{id =2}), one healthy, censored patient
#  (\code{id =3}) and one ill, censored patient (\code{id =4})
msm_data_example <- data.frame(id = c(1,1,2,2,2,3,3,4,4,4), Tstart = c(0,0,0,0,0.5,0,0,0,0,1),
                               Tstop = c(0.8,0.8,0.5,0.5,1.2,1.3,1.3,1,1,1.1),
                               duration = c(0.8,0.8,0.5,0.5,0.7,1.3,1.3,1,1,0.1),
                               from = c(1,1,1,1,2,1,1,1,1,2), to = c(2,3,2,3,3,2,3,2,3,3),
                               status = c(0,1,1,0,1,0,0,1,0,0), trans = c(1,2,1,2,3,1,2,1,2,3),
                               recruitment_date = c(0,0,0.3,0.3,0.3,0.7,0.7,0.9,0.9,0.9))
fit_tihmm(msm_data_example, transition_matrix = tmat_example)

}
