% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cum_haz_fct_exp}
\alias{get_cum_haz_fct_exp}
\title{Helper function to create cumulative hazard function for exponential distribution}
\usage{
get_cum_haz_fct_exp(rate)
}
\arguments{
\item{rate}{Rate of exponential distribution}
}
\value{
conditional hazard function
}
\description{
Helper function to create cumulative hazard function for exponential distribution
}
\examples{
rate <- 1
reference_ch_fct <- mvoslr:::get_cum_haz_fct_exp(rate)
}
\keyword{internal}
