% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cum_haz_fct_weibull}
\alias{get_cum_haz_fct_weibull}
\title{Helper function to create cumulative hazard function for exponential distribution}
\usage{
get_cum_haz_fct_weibull(parameters)
}
\arguments{
\item{parameters}{Vector of length two containing shape and scale parameter}
}
\value{
Conditional hazard function
}
\description{
Helper function to create cumulative hazard function for exponential distribution
}
\examples{
shape <- 1.5
scale <- 1
reference_ch_fct <- mvoslr:::get_cum_haz_fct_weibull(c(shape, scale))
}
\keyword{internal}
