% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse_normal_combine}
\alias{inverse_normal_combine}
\title{Inverse normal combination function with arbitrary weights}
\usage{
inverse_normal_combine(p, weights)
}
\arguments{
\item{p}{Vector of staagewiswe p-values}

\item{weights}{Vector of weights}
}
\value{
Overall p-value for vector of stagewise p-values with given weights
}
\description{
Inverse normal combination function with arbitrary weights
}
\examples{
my_p <- c(0.2, 0.1, 0.3)
my_weights <- rep(1/sqrt(4), 4)
mvoslr:::inverse_normal_combine(my_p, my_weights)

}
\keyword{internal}
