% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_smm.R
\name{log_lik_single_obs_smm}
\alias{log_lik_single_obs_smm}
\title{Helper function to set up log likelihood function to estimate parameters in Semi-Markov model}
\usage{
log_lik_single_obs_smm(single_obs_data, shape, scales)
}
\arguments{
\item{single_obs_data}{single row of data from multi-state model, column names}

\item{shape}{joint shape parameter of transition intensity functions}

\item{scales}{vector of scale parameters of transition intensity functions}
}
\value{
log likelihood for single observation
}
\description{
Helper function to set up log likelihood function to estimate parameters in Semi-Markov model
}
\examples{
msm_data_example <- data.frame(id = c(1,1,2,2,2,3,3,4,4,4), Tstart = c(0,0,0,0,0.5,0,0,0,0,1),
                               Tstop = c(0.8,0.8,0.5,0.5,1.2,1.3,1.3,1,1,1.1),
                               duration = c(0.8,0.8,0.5,0.5,0.7,1.3,1.3,1,1,0.1),
                               from = c(1,1,1,1,2,1,1,1,1,2), to = c(2,3,2,3,3,2,3,2,3,3),
                               status = c(0,1,1,0,1,0,0,1,0,0), trans = c(1,2,1,2,3,1,2,1,2,3),
                               recruitment_date = c(0,0,0.3,0.3,0.3,0.7,0.7,0.9,0.9,0.9))
log_lik_single_obs_smm(msm_data_example[1,], shape = 1.5, scales = c(0.4, 0.2, 0.5))
}
\keyword{internal}
