% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msm_to_trial_data}
\alias{msm_to_trial_data}
\title{Transform raw multi-state model data to trial data using simple censoring mechanism}
\usage{
msm_to_trial_data(msm_data, accrual_duration, follow_up_duration)
}
\arguments{
\item{msm_data}{Data frame containing information about any (possible) transition with (at least) the following columns:
\itemize{
\item id - ID of the patient to whom this entry belongs (requires continuous numbering starting with 1)
\item Tstart - Start of the observation period for this transition
\item Tstop - Stop of the observation period for this transition
\item status - Transition could (1) or could not be (0) observed
\item (recruitment_date) - If a recruitment date is supplied, it will be used to create corresponding censoring
}}

\item{accrual_duration}{Accrual duration of the trial to be simulated}

\item{follow_up_duration}{Duration of the follow-up period of the trial to be simulated}
}
\value{
Data frame with results of simulation
}
\description{
Transform raw multi-state model data to trial data using simple censoring mechanism
}
\examples{
raw_msm_data_example <- data.frame(id = c(1,1,2,2,2,3,3,4,4,4), Tstart = c(0,0,0,0,0.5,0,0,0,0,1),
                                   Tstop = c(0.8,0.8,0.5,0.5,1.2,1.3,1.3,1,1,1.1),
                                   status = c(0,1,1,0,1,0,0,1,0,0))
msm_to_trial_data(raw_msm_data_example, accrual_duration = 1, follow_up_duration = 0.5)
}
\keyword{internal}
