% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_enhanced_result.R
\name{new_mvoslr_enhanced_result}
\alias{new_mvoslr_enhanced_result}
\title{Constructor for class "mvoslr_result" which is used to save and pass results}
\usage{
new_mvoslr_enhanced_result(
  raw_process = array(),
  stagewise_p_values = matrix(),
  rejection_stage = integer(),
  remaining_analyses = integer(),
  vector_norm = character(),
  variable_parameter = character(),
  parameter_values = numeric(),
  follow_up_fixed = logical()
)
}
\arguments{
\item{raw_process}{Values of multivariate process at analysis dates}

\item{stagewise_p_values}{Stagewise p-values}

\item{rejection_stage}{Number of stage at which the null hypothesis was rejected (is "NA" if null hypothesis could not be rejected)}

\item{covariation_matrices}{Estimated covariance matrix for multivariate process at analysis date}

\item{multivariate_test_statistics}{Stagewise multivariate, standardised test-statistics}

\item{univariate_test_statistic}{Stagewise univariate test-statistics}
}
\value{
Object of class "mvoslr_result"
}
\description{
Constructor for class "mvoslr_result" which is used to save and pass results
}
\keyword{internal}
