% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_power.R
\name{new_mvoslr_power_object}
\alias{new_mvoslr_power_object}
\title{Constructor for class "mvoslr_power_object" which is used to save and pass results of power simulations.
If simulations for several values of a design parameter are conducted, results for each value are saved here.}
\usage{
new_mvoslr_power_object(
  power = numeric(),
  rejection_stages = matrix(),
  means = array(),
  variances = array(),
  runs = numeric(),
  variable_parameter = character(),
  follow_up_fixed = logical(),
  parameter_values = numeric()
)
}
\arguments{
\item{power}{Overall simulated power}

\item{rejection_stages}{Relative frequency of rejections of the null hypothesis in each stage}

\item{means}{Mean value of the multivariate process at analysis dates}

\item{variances}{Mean estimated covariance matrices of the multivariate process at analysis dates}

\item{runs}{Number of simulation runs}

\item{variable_parameter}{Varied parameter (if any)}

\item{follow_up_fixed}{Boolean variable indicating if the follow-up duration is held fixed when the accrual duration ("a") is varied}

\item{parameter_values}{Values of the parameter which is varied}
}
\value{
Object of class "mvoslr_result"
}
\description{
Constructor for class "mvoslr_power_object" which is used to save and pass results of power simulations.
If simulations for several values of a design parameter are conducted, results for each value are saved here.
}
\keyword{internal}
