% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_model.R
\name{new_reference_model}
\alias{new_reference_model}
\title{Constructor for class "reference_model"}
\usage{
new_reference_model(
  transition_matrix = matrix(),
  intensities = list(),
  type = character(),
  parametric = FALSE,
  parameters = matrix()
)
}
\arguments{
\item{transition_matrix}{Matrix of transitions between states as in mstate package}

\item{intensities}{Cumulative hazard functions for transitions in this model}

\item{type}{Reference multi-state model is either Markov (\code{model_type = "M"}) or Semi-Markov (\code{model_type = "SM"})}

\item{parametric}{Cumulative intensity functions can be parametric with Weibull form}

\item{parameters}{If cumulative intensity functions are parametric, parameters can be supplied here with one column for each transition, shape parameters in first and scale parameters in second row}
}
\value{
Object of class "reference model" to be used in planning and execution of multivariate one-sample log-rank test
}
\description{
Constructor for class "reference_model"
}
\examples{
library(mstate)
tmat_example <- transMat(x = list(c(2,3),c(3),c()), names = c("a", "b", "c"))
model_type_example <- "SM"
cumhaz_12_example <- function(t) t^1.1
cumhaz_13_example <- function(t) t^1.2
cumhaz_23_example <- function(t) t^0.9
cumhaz_example <- list(cumhaz_12_example, cumhaz_13_example, cumhaz_23_example)
parametric_example <- TRUE
parameters_example <- matrix(c(1.1, 1, 1.2, 1, 0.9, 1), nrow = 2,
                             dimnames = list(c("shape", "scale"), 1:3))
model_example <- new_reference_model(transition_matrix = tmat_example,
                                     intensities = cumhaz_example,
                                     type = model_type_example, parametric = parametric_example,
                                     parameters = parameters_example)
}
