% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_model.R
\name{plot.reference_model}
\alias{plot.reference_model}
\title{Plot method for an object of class "reference_model". Plots cumulative intensity functions of all transitions.}
\usage{
\method{plot}{reference_model}(model, ...)
}
\arguments{
\item{model}{Object of class "reference_model"}
}
\value{
Returns the object itself
}
\description{
Plot method for an object of class "reference_model". Plots cumulative intensity functions of all transitions.
}
\keyword{internal}
