% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_result.R
\name{print.mvoslr_result}
\alias{print.mvoslr_result}
\title{Print method for an object of class "mvoslr_result"}
\usage{
\method{print}{mvoslr_result}(result)
}
\arguments{
\item{result}{Object of class "mvoslr_result"}
}
\value{
Returns the object itself
}
\description{
Print method for an object of class "mvoslr_result"
}
\keyword{internal}
