% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_model.R
\name{reference_model_weibull}
\alias{reference_model_weibull}
\title{Helper function to create object of class "reference_model" from parameters of transition intensity functions with Weibull form.}
\usage{
reference_model_weibull(
  transition_matrix = matrix(),
  type = character(),
  parameters = matrix()
)
}
\arguments{
\item{transition_matrix}{Matrix of transitions between states as in mstate package}

\item{type}{Reference multi-state model is either Markov (\code{model_type = "M"}) or Semi-Markov (\code{model_type = "SM"})}

\item{parameters}{Matrix of parameters for Weibull-shaped transition intensities with one column for each transition, shape parameters in first and scale parameters in second row}
}
\value{
Object of class "reference model" to be used in planning and execution of multivariate one-sample log-rank test
}
\description{
Helper function to create object of class "reference_model" from parameters of transition intensity functions with Weibull form.
}
\examples{
library(mstate)
tmat_example <- transMat(x = list(c(2,3),c(3),c()), names = c("a", "b", "c"))
model_type_example <- "SM"
parameters_example <- matrix(c(1.1, 1, 1.2, 1, 0.9, 1), nrow = 2,
                             dimnames = list(c("shape", "scale"), 1:3))
model_example <- reference_model_weibull(transition_matrix = tmat_example,
                                         type = model_type_example,
                                         parameters = parameters_example)
}
