% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_msm}
\alias{simulate_msm}
\title{Wrapper function for mstate::mssample with some adjustments}
\usage{
simulate_msm(transition_matrix, model_type, cum_hazards_frame, sample_size)
}
\arguments{
\item{transition_matrix}{Matrix of transitions between states as in mstate package}

\item{model_type}{Multi-state model is either Markov (\code{model_type = "M"}) or Semi-Markov (\code{model_type = "SM"})}

\item{cum_hazards_frame}{Data frame containing values of hazard functions of each transition}

\item{sample_size}{Number of subjects to simulate}
}
\value{
Data frame with results of simulation
}
\description{
Wrapper function for mstate::mssample with some adjustments
}
\examples{
tmat_example <- transMat(x = list(c(2,3),c(3),c()), names = c("a", "b", "c"))
model_type_example <- "SM"
cumhaz_12_example <- function(t) t^1.1
cumhaz_13_example <- function(t) t^1.2
cumhaz_23_example <- function(t) t^0.9
cum_hazards_example <- list(cumhaz_12_example, cumhaz_13_example, cumhaz_23_example)
disc_cum_hazards_frame <- mvoslr:::discretize_functions(cum_hazards_example,
                                                        max_argument = 10, time_steps = 1000)
sample_size_example <- 100
mvoslr:::simulate_msm(transition_matrix = tmat_example, model_type = "M",
                      cum_hazards_frame = disc_cum_hazards_frame, sample_size = 10)
}
\keyword{internal}
