% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_enhanced_result.R
\name{summary.mvoslr_enhanced_result}
\alias{summary.mvoslr_enhanced_result}
\title{Summary method for an object of class "mvoslr_result"}
\usage{
\method{summary}{mvoslr_enhanced_result}(result)
}
\arguments{
\item{result}{Object of class "mvoslr_result"}
}
\value{
Returns the object itself
}
\description{
Summary method for an object of class "mvoslr_result"
}
\keyword{internal}
