% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoslr_power.R
\name{summary.mvoslr_power_object}
\alias{summary.mvoslr_power_object}
\title{Summary method for an object of class "mvoslr_power_object"}
\usage{
\method{summary}{mvoslr_power_object}(power_obj)
}
\arguments{
\item{power_obj}{Object of class "mvoslr_power_object"}
}
\value{
Returns the object itself
}
\description{
Summary method for an object of class "mvoslr_power_object"
}
\keyword{internal}
