% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_model.R
\name{summary.reference_model}
\alias{summary.reference_model}
\title{Summary method for an object of class "reference_model"}
\usage{
\method{summary}{reference_model}(model)
}
\arguments{
\item{model}{Object of class "reference_model"}
}
\value{
Returns the object itself
}
\description{
Summary method for an object of class "reference_model"
}
\keyword{internal}
