% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdata_to_df.R
\name{msdata_to_df}
\alias{msdata_to_df}
\title{Transform data of type msdata from mstate package to a data frame}
\usage{
msdata_to_df(msdata)
}
\arguments{
\item{msdata}{}
}
\value{
data frame with column names corresponding to the columns names of msdata objects
}
\description{
Transform data of type msdata from mstate package to a data frame
}
