% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msdata_to_df}
\alias{msdata_to_df}
\title{Transform data of type msdata from mstate package to a data frame}
\usage{
msdata_to_df(msdata)
}
\arguments{
\item{msdata}{Data set of type msdata from \code{mstate} package}
}
\value{
data frame with column names corresponding to the columns names of msdata objects
}
\description{
Transform data of type msdata from mstate package to a data frame
}
\examples{
library(mstate)
tmat_example <- transMat(x = list(c(2,3),c(3),c()), names = c("a", "b", "c"))
model_type_example <- "SM"
cumhaz_12_example <- function(t) t^1.1
cumhaz_13_example <- function(t) t^1.2
cumhaz_23_example <- function(t) t^0.9
cum_hazards_example <- list(cumhaz_12_example, cumhaz_13_example, cumhaz_23_example)
disc_cum_hazards_frame <- mvoslr:::discretize_functions(cum_hazards_example,
                                                        max_argument = 10, time_steps = 1000)
sim_data <- mstate::mssample(Haz = disc_cum_hazards_frame,
                             trans = tmat_example,
                             M = 10,
                             output = "data")
sim_frame <- mvoslr:::msdata_to_df(sim_data)
}
\keyword{internal}
